requirements.txt
Pillow==10.0.1
Pillow==10.0.1
python-minifier==2.9.0
